Use make-gui-namespace as initial namespace; fixes #555
This is a minimal fix for issue #555, while not breaking programs that
use racket/gui/base.

However make-gui-namespace is a little too much for the case where we
we actually use that in the REPL (i.e. a plain racket-repl command
where we don't module->namesapce) because requires racket/class and
racket/gui/base. Really the REPL in that case should only have
racket/base.

The next commit will refine that, as well as take the opportunity
reorganize some code, which I think will be helpful -- but which may
also obscure the gist of the required change, ergo this interim
minimal commit.