Refine previous commit's fix for #555 and reorganize some code
In the gui case, make the initial namespace more limited: It is
make-base-namespace with racket/gui/base merely _attached_. By
contrast, in the previous commit, make-gui-namespace also _requires_
racket/class and racket/gui/base. Although moot in the case where the
initial namespace is only used to module->namespace a specific file,
it matters in the case where a user does a simple racket-repl (not any
specific file); that is supposed to be just racket/base, until/unless
the user requires more in the REPL.

Also, instead of using a parameterize form around the thread creation,
directly set most parameters inside the REPL thread thunk itself. I
think the resulting proximity and order in the source makes it more
clear what is the order of execution and effective scope of the
parameters.