Fewer, documented dependencies
Eliminate dependency on gui-debugger/marks via "vendoring".

Make macro-debugger-text-lib a "soft" dependency; that is, the back
end can still start, but if the racket-{trim base}-requires commands
are used they'll give an error message.

Note that html-lib used to be a dependency until the very recent
commit 46df4d9.

As a guide to people wanting to use minimal Racket, list the remaining
dependencies in racket-mode.org.

Of these, drracket-tool-lib might soon be able to become a new
drracket-tool-text-lib -- not depending on gui-lib -- at least for
versions of Racket new enough to use the new package. See:

  https://github.com/racket/drracket/issues/517