Change handling of user-name and ssh-port
Default these to nil. Furthermore, in doc string for
racket-back-end-function encourage users to default these to nil,
unless they would actually supply them in a tramp file name.
For example, leave them nil if they are specified in ~/.ssh/config.
Also leave ssh-port nil for the default value of 22.

The idea here is to improve the equivalence of tramp file names
created by racket-file-name-back-to-front with buffer file names.