Preserve current-get-interaction-input-port; fixes #562
When we eagerly load racket/gui/base, it sets
current-get-interaction-input-port to a function that wraps the
original. The wrapper filters the original input port to call yield
when the REPL is waiting for input, so that gui program event loops
can run. Therefore it is important for us not to change the value of
current-get-interaction-input-port.

For issues like #519 and #556, we do need to port-count-lines! on the
input port accepting the user's REPL input. Doing port-count-lines! on
the TCP input port upon making a REPL connection is not sufficient for
input in the read-eval-print-loop thread. For that, we need to do
port-count-lines! on the value ((current-get-interaction-input-port))
in our get-interaction function, which is called from our value for
the current-prompt-read parameter.

(I had actually used this approach for a commit 53bfb97, still visible
on GitHub's thread for issue #562. But I had a brain fart when
deciding to "simplify" that in the eventually merged commit 4dbf807.)