Mitigate Windows Emacs certificate problem
Add a deps-windows-ci target to Makefile and use it from test.yml for
Windows CI.

This is like the deps target, but sets the Emacs variable
network-security-level to 'low to ignore certificate check failures
when doing package-refresh-contents.

This is a hack that should be reverted as soon as the situation with
Windows Emacs and Let's Encrypt certificates is resolved.