Remove end user dependency on Emacs package faceup
Move racket-tests.el to a test/ subdir so that it is no longer
delivered in a MELPA package or compiled.

As a result, we no longer need faceup -- which is only used for
running tests -- to be in the end user Package-Requires list.

Note that `make deps` DOES still install faceup for CI and for people
hacking on Racket Mode.