Attach same instance of file/convertible to user namespaces
Very recently commit f89d552 fixed an issue with handling images files
from remote back ends. I know it fixed it because I saw it working a
few days ago!

Today: I experienced it consistently still working with a local back
end --- but consistently /not/ working with a remote back end.
Basically, file/convertible convertible? was returning false for pict
objects. And yet, file/convertible works fine on the remote with
example programs in command-line racket. The only difference this time
is that the Racket install there is minimal; no gui-lib.

This reminded me of an issue in Racket Mode many years ago, which
seemed to be about file/convertible using struct properties, but
Racket structs are generative, so sometimes it would not work.

Back then I used dynamic-require on image.rkt, as in commit 51e48dd.

But later, this no longer seemed necessary. (Would need to search the
commit history much more to find when/why.)

This commit uses what I think is a simpler/clearer approach: Attach
the same instance of file/convertible to the namespace initially
created for end user programs.

TL;DR: Maybe when racket/gui/base is eagerly loaded (when I was using
racket with gui-lib installed), and we attach that instance to user
namespaces, a single instance of file/convertible is attached, too --
maybe it comes along for the ride?? Anyway, let's just always attach
file/convertible to be sure in all scenarios.