Use two steps to get from buffers to back ends
A first design had back end property lists baked into a buffer-local
variable `racket-back-end'. This made it tricky to change the
configuration, because all the buffers had to be refreshed.

To avoid this, the next design called a `racket-back-end` function.
Although not "stale", it was "too fresh" -- too much work done too
often.

This design tries to be the warm bowl of oatmeal, using two steps:

1. A `racket-back-end-name' function determines the _name_ of a back
end that the buffer should use. This has a user hook to handle the
case of more than one back end per host. This may be called often, but
user hooks will typically just (string-match-p "^/some/path"
default-directory), which is fast.

2. A `racket-back-end` function takes the name and returns the plist
configuration info. Similarly there are trivial functions to return
the name of the back end's process buffer.

Furthermore, this tries to optimize the configuration experience, so
that users only need to specify properties whose default values they
want to change. Not only is this more convenient it is less
error-prone.

As before, in common scenarios users need not even specify any
configuration. Back end configs are created automatically for hosts
they user hasn't specified, and will usually "just work". (At least
that's the goal.)

Also:

- Change racket--winp to be boolean (not just truthy).

- Rewrite doc string for racket--buffer-file-name.