Yet another approach to configuration
The primary configuration item for a back end is now a "directory",
which is used for several purposes:

- As a unique id for the back end (the role that :name used to play).

- As a pattern to match against default-directory, to decide if a
back end should be used for a buffer.

- As the way to connection properties (host and maybe user and port)
when the directory is file-remote-p.

So e.g.:

 - "/path/to/project" matches local files at/under that dir.

 - "/" matches all other local files.

 - "/ssh:user@host:/path/to/project" matches remote files at "host"
   at/under /path/to/project

 - "/ssh:user@host:" matches all other remote files at "host"

There are still additional keyword properties for special cases, but
often the directory will suffice.

I think this is definitely cleaner/simpler for an end user to
configure. Plus as before, manual configuration usually won't be
needed for the case of one back end per host: If no existing
configuration matches, one will automatically be created and
used, and it will usually "just work".