Use racket--escape-string-or-comment in racket--calculate-indent
This is to find the initial "orientation point".

Fixes #373.
Fixes #588.

Although I was nervous that the use of syntax-ppss might make it
extraordinarily slow, such as near the end of very long files, it does
not seem to be when doing individual line indents.

(Although it is somewhat slower doing indent-region on entire files,
it's on the order of single digit seconds for such very long files.
Someday if re-indenting entire files becomes an important use
case (currently it's not), we should start by providing an
indent-region-function, and optimizing from there.)