Review/change use of string-match and match data
Many of our uses of `string-match' could and should instead be uses of
`string-match-p'.

The remainder should be careful to use `save-match-data'.

Nothing should use the local variable `inhibit-changing-match-data'
because it becomes obsolete in a future version of Emacs:

  http://git.savannah.gnu.org/cgit/emacs.git/commit/etc/NEWS?id=a26824e5cbc454a3614b32d131bbdd74f3c03735

Although we weren't using `inhibit-changing-match-data', this news
item was the motivation for this commit.