Fix uses of define-obsolete-face-alias
I misunderstood the second value to be a doc string but instead it is
supposed to be a face symbol.

This has been wrong for racket-paren-face for many years (!).

How I noticed it was for the deprecation of racket-selfeval-face in
very recent commit 991f98c. Today I was changing among Emacs themes
and got a symbolp error wrt to the string.