pretty-print-size-hook must not return inexact
Although we use ceiling, this could result in a length like 35.0, and
the innards of pretty-print will error when this value is used as the
length argument of make-string.

Although this can happen when pretty-printing a list of picts -- I
noticed it today while using pict to generate front/back end
architecture images for the documentation -- I've never seen it happen
before when pretty-printing other lists of images over the years.