racket--describe-goto: Move past multiple anchors for same text
Motivating example:

  https://docs.racket-lang.org/web-server/http.html#%28part._bindings%29

has two anchors, which end up in the buffer as text with a 'display
property value of "". That is, normal user navigation will skip over
these; effectively they'll be "invisible".

If we racket-describe-search for "Bindings", we get the first anchor.
The old behavior was we'd just move point one char forward --- but
that left point on the second anchor. We want to move point past that,
too, to be on the "B".

One practical benefit is that, if user does another
racket-describe-search, the default at the prompt -- which is found
using (thing-at-point 'symbol) -- will be "Bindings" not nil. This was
confusing because it made it seem initially that "Bindings" wasn't in
the completions list.