Indent Typed Racket let variants; fixes #569
Also properly indent a for or for* form with an unnecessary ": Void"
type annotation.

The desired indentation here is not quite the "classic Lisp" style
with respect to distinguished forms that appear on their own line (not
on the first line). The classic indent is twice the body indent (e.g.
4 instead of 2). But the desired indent for some Racket forms is to
align the other distinguished forms with the first distinguished form.
This style was introduced (AFAIK) with Racket's for/fold form -- you
want the lists of accumulators and for-clauses to be indented the
same. In this style the distinguished forms are indented like
arguments in a function application (with the following,
non-distinguished forms getting normal body indent).

To support this have racket--indent-special-form accept a negative
number. The absolute value is the number of distinguished forms.
Positive is classic indent, negative is the other style.

Given this we no longer need racket--indent-for/fold-untyped -- it was
just doing this other style in an ad hoc and needlessly complicated
way.

(Note: The "negative number of distinguished forms" thing is for now a
private implementation detail. I'm not confident it's a generally good
idea. And I'm definitely not sure that "negative number" is the best
"API" for end user indent spec property values. So for now I'm not
documenting this for end users.)

Also update some comments.