racket-show: Add optional argument transient-p
When true this sets a pre-command-hook to (racket-show "") before the
next command runs. This is added for possible use by things like #573,
in contrast to the sole status quo use in racket-xp-mode.

Also add argument type checks.

Also take the opportunity to rename the first arguments from "v" to
"str" to emphasize that if non-nil they must be strings. (In the
racket-show-echo-area function we use (message "%s" str) not because
the value is allowed to be a non-string, but instead to make sure that
any % characters in the string are treated literally.)