Fix bug introduced in commit a093a44 for #569; fixes #574
The determination of the column of the first distinguished form,
although necessary for the new "negative" method, was moving point and
causing a problem in some scenarios.

Also fix a mistake in the doc string.

Also rename the "pos" variable to "count" (it has nothing to do with
buffer positions; it's a count of forms).