Simplify run command "maybe-mod" handling
Remove unnecessary code, some dating back to the run command being an
xrepl style command that had to deal with arbitrary user input.

These days the back end run command accepts a "what" argument from the
Emacs front end that is one of:

  (complete-path-string . submod-symbols)
  nil

We can change that to a valid module-path? that is one of:

  (and/c path? complete-path?)
  ('submod (and/c path? absolute-path?) . submod-symbols)
  #f

(Note that although module-path? disallows complete path _strings_
unless wrapped in (file _), it allows complete path _objects_ and
indeed we require them to be complete. Remember "complete" ~=
"absolute plus a drive letter when on Windows".)

With that simpler subset of module-path? we can just define a couple
helper functions to extract the file and to make a prompt string.

We no longer need a "mod" struct and a bunch of code to handle all
variants of module-path?.