WIP commit
Change the token-map struct into a hash-lang% class. We need to supply
an object% with certain methods for drracket:indentation functions,
so, we may as well just make the whole thing a class instead of a
struct. (Maybe this could be refactored into smaller pieces via mixins
or something because it's starting to feel a little "fat".)

Abandon our old strategies for parens and whitespace. Just go with the
latest flow from drracket/expeditor.