Use syntax-color/color-textoid when available
Handle it being available (with new-enough versions of Racket and/or
syntax-color-lib) or not. After all, a goal of Racket Mode is for it
work to the extent possible on older versions of Racket.

This is a good opportunity to move the tests to their own file, since
they need to be skipped when color-textoid is not available. Anyway,
they were starting to become unwieldy in the same file; it was
becoming more tests than tested.

Some of this will become N/A when we move it to syntax-color-lib
itself; then hash-lang-bridge.rkt will need to do a similar test. But
moving the tests to their own file is good prep, as those will end up
in syntax-color-test not syntax-color-lib.