Remove {beginning end}-of-line; somewhat optimize paragraph numbers
From https://github.com/racket/expeditor/issues/10 it sounds like the
"traditional" approach isn't desirable after all; remove it.

Make the paragraph numbering "moderately" optimized: On do-udpate! we
invalidate it, and then the paragraph methods recalculate it
on-demand.

This is a mid point on design spectrum between "naively scan from 0
every single time" and "do some minimal rebuild on every update".

Note that I'm not 100% confident about the concurrency safety. But I'm
committing anyway, for now, because I plan to take hard look at
concurrency for the class, soon, including this as well as the
potential for update generations arriving on command threads out of
order and needing to be queued something like TCP packets.