Test our speed vs. racket:text%, when used for range-indent
This test allows ours to take <= 3X the time.

Although I'm not sure that's good enough, it sets an initial bar.

(Although 3X isn't superb, the test example is range-indents
rhombus/demo.rkt, a 600 line file -- and does so 10 times. The
total time is around 54 ms for us vs. 24 ms racket:text%.)

(To some extent this is measuring how "fancy" we are being in
supporting the position-paragraph method, which effectively is
position->line-number.)