Start to use token-tree%, and add open-input-text
This is a checkpoint commit where not all tests pass; those that fail
do so because (for now) we are always re-tokenizing from beginning to
end.

However it works well enough to run the benchmarks. Now we are about
2X color-text%, whereas before with interval-map access we were
10-15X. Although I want to figure out why 2X, at least it's much
better.

This commit changes to use 0-based positions throughout. This seems
more straightforward to use with token-tree%, plus it means we now
longer have to keep it straight with the color-textoid% portion which
uses 1-based positions. It will mean that hash-lang-bridge.rkt will
need to covert 1 <-> 0, but that's one simple "edge" place to do so.

This commit also adds open-input-text to text-lines; it is similar to
that from snipfile.rkt, but of course simpler because we needn't deal
with snips.