Use syntax-color/racket-{indent navigation}
Also: No longer use paren-matches and quote-matches to apply syntax,
especially since the comment-matches proposal was rejected. If that
means performance and/or Emacs ecosystem compatibility is not as good,
so be it. The intent will be that you use racket-hash-lang-mode if
your highest priority is fidelity to the hash-lang mechanism.