New approach using font-lock-fontify-region-function
Use Emacs font-lock machinery to drive the updating.

In particular this means that the hash-lang% update! no longer
pushes/streams all changed tokens. For large files that was very slow
-- not just for the initial update, but for certain updates near the
start of the file, such as adding/removing an unmatched " or, even
worse, in #lang scribble/manual, an unmatched @(.

Instead: Now the hash-lang% update! methods emits a single (list
'invalidate gen beg end) notification to describe the span where
tokens changed. The front end can use this to call (font-lock-flush
beg end), i.e. "please re-fontify this region if/as/when needed". Note
that doesn't mean font-lock will eagerly happen for any/all of that
region. It will happen only as/when any of that region needs to be
displayed, e.g. due to scrolling.

Our value for font-lock-fontify-region-function (which is called only
when some region needs to be fontified) uses the hash-lang% get-tokens
method to fetch tokens for that region and apply the face properties.
Furthermore this is an async command, so it doesn't block Emacs if the
tokens aren't yet ready.

AFAICT this is sufficient/correct for applying face properties, to
make things appear correctly.

I'm still not 100% crisp on how/when/why we should apply syntax-table
properties. Technically, maybe that should be done via
syntax-ppss-flush-cache and a value for syntax-propertize-function?
But I'm not sure how/when that matters, given that we now rely
entirely on drracket:grouping-position (defaulting to
racket-grouping-position), and those work in terms of the token and
paren trees on the back end, not in terms of Emacs char syntax. As for
other Emacs features and packages, those should be fine because when
we know the grouping-position is racket-grouping-position (for
s-expression langs), we use a standard syntax-table. Those should
continue to work as well as before (not perfectly compared to a lexer
handling every little edge case, but no worse than they have worked
all along). And when a lang does NOT use racket-grouping-position, we
use a plain/bare syntax-table that truly assumes nothing, i.e. no
chars signify parens or strings or comments.