Fix initial lexer mode; rename our token-tree% struct
Fix determination of initial position from which to re-tokenize and
the initial lexer mode from which to use. Use all backup amounts.
Also, the initial lexer mode to use is the one stored with the
previous token. This matters e.g. for at-exp langs.

Also, rename the struct we store in token-tree% nodes from "token" to
"data", and remove the "paren" member which we've no longer been using
since we started using paren-tree%.