Use on-changed-xxx virtual methods instead of on-notify proc
If we're using racket/class we may as well "when in Rome" this.

Also, change the "on new lang values" notification to be "empty" and
let the overrider/receiver query whatever they might be interested in.

Also update hash-lang.el for this, as well as the fact that there is
always a line-indenter (if a given lang doesn't supply one then the
racket indenter is used). Only a range-indenter is optional.