Fix contig-same-count and handle dont-stop
What we were actually doing: Counting the number of tokens that were
the same, even if intervening tokens differed.

What we wanted, as reflected by the variable name: Stopping when we
have N same tokens in a row.

Do that.

At the same time, handle a lexer returning a `dont-stop` struct, and
don't stop early when that is true. (I'm not sure if we should
completely reset contig-same-count back to zero; the docs for
dont-stop seem not to require that. At any rate, this commit will at
least prevent us from barfing on a dont-stop value, and, we will keep
going so long as the lexer keeps returning that.)