Support hash-lang in "tail" of REPL buffers
* Back end

Group the lang-supplied values in a new `lang-info` struct.

Extend hash-lang% with an optional `other-lang-source` init-field --
read-language from that instead of the content.

Have hash-lang-bridge.rkt accept an optional REPL session ID, and use
that to get the path and source text for `other-lang-source`.

* Front end

Change the hash-lang IDs (used by hash-lang-bridge.rkt to map
serializable IDs like strings to hash-lang% object instances) to a
gensym stored in a new buffer-local variable, racket--hash-lang-id.

Add new buffer-local variable racket--hash-lang-offset, which is the
offset into a buffer that the hash-lang applies to. When we're
enhancing a racket-mode edit buffer, this is simply 0. When we're
enhancing a racket-repl-mode buffer, this is set to the start of user
input, i.e. after the last output field, i.e. just the "current
interaction". We use the hash-lang for font-lock and indent only after
that offset.

Use racket--{before after}-run-hook values to automatically
disable/enable racket-hash-lang-mode for a racket-repl-mode buffer if
its corresponding edit buffer is using racket-hash-lang-mode. In other
words, this is /not/ something a user configures in a
racket-repl-mode-hook. (It is more similar to racket-debug-break-mode,
something that we enable/disable automatically.)

* More explanation

Although it is possible users might configure to use one REPL buffer
per edit buffer, similar to Dr Racket, that's not required. Users may
share a REPL buffer multiple edit buffers. In that case each
`racket-run' "takes turns" using a REPL buffer; the buffer may
accumulate I/O associated with multiple edit buffers, each of which
could be using a different #lang.

Furthermore, a REPL buffer grow very large, including evaluation
results with large printed representations, as well as user programs
displaying a large amount of output. As a result even our "classic"
font-lock for racket-repl-mode can sometimes be inappropriate or slow.
It would be even slower for racket-hash-lang-mode, where we need to
talk to the back end.

Therefore we take the approach of only handling the "tail" of the REPL
buffer.

I'm not yet sure this approach is rigorous or sufficient, but I'm
somewhat confident that something in this ballpark is going to turn
out to be how we handle it.