racket--describe-goto: set-window-point not goto-char; fixes #590
I think this is unique to racket-describe-mode due to how we create
the buffer and (maybe) window, to handle the case where the buffer
will be used by company-mode. In some circumstances the buffer's
window might not be selected, in which case goto-char won't be
sufficient to move point in the window.