Supply dedicated support for :company-doc-buffer; fixes #592
Once upon a time, racket-describe-mode showed just a "peephole" into a
documentation page for the item in question. This meant it was also
generally suitable for use by :company-doc-buffer. And we used it for
both purposes.

Later it evolved to show the full documentation page, much like how
you'd view the docs using a "normal" external web browser. This meant
it was so suitable for use by :company-doc-buffer (especially when
something special like company-quickhelp-mode is being used, which
tries to show it all in a pos-tip frame).

Furthermore it evolved to where the buffer returned by
racket--do-describe wouldn't even contain the content, yet; it was
being fetched asynchronously. This meant it is was very unsuitable for
use by :company-doc-buffer (again, especially wrt company-quickhelp
which just immediately grabs the buffer text).

Instead:

We share less code, and for :company-doc-buffer we restore the
"peephole" approach and make it synchronous.

We do still use shr to handle all the edge cases of changing an HTML
DOM into formatted plain text. Unlike racket-describe, however, we:

- Set shr-width to 76, to accommodate company-quickhelp-mode width of
80.

- Set shr-external-rendering-functions to use racket-render-tag-span,
only, to get some fancy formatting. (We don't bother with the others,
which are mostly applicable only to racket-describe's full doc pages.)

- Always recreate buffer; otherwise sometimes shr-insert will garble
things, e.g. for the `match` form the bluebox will end up at the end
not the beginning.