add direct requires to keep native compiler happy
In emacs 28+ the native compiler fails if a function is used without being
defined, being in a direct import, or sometimes having a direct import and
a declare-function expression associated with it (not entirely clear why
the compiler needs declare-function in some cases where the funciton is
provided by a direct import but not in other cases).

If these fixes are not in place make compile will fail if run with emacs
built with native compile.

Some docstring formatting changes were necessary to keep the compiler happy.