Debug breakable positions: Use racket-file-name-back-to-front; fixes #597
The back end returns a list of breakable positions which are (cons
path positions). The path needs back-to-front translation, as with any
path from the back end.

The original issue #597 reported this for Windows (where Racket uses
native paths with back slashes whereas Emacs uses forward slashes for
file names and therefore names of buffers visiting files).

However this is also an issue for remote back ends. That is actually
the scenario with which I tested this fix. Although I'm being a bit
lazy by committing without testing this on Windows, too, I'm fairly
confident. If the original reporter doesn't confirm the fix works
there, I'll dig in deeper; meanwhile this definitely does fix a
problem with remote back ends.