racket-debug-run-to-here: File name needs front-to-back; fixes #597
Be sure to use racket-file-name-front-to-back on the file name given
to the back end by racket-debug-run-to-here.

Also, although not part of the original bug report, make sure the
breakable positions are sorted. Once upon a time the back end might
have done this but it certainly doesn't today; do so in the front end.