Display faded s-expression comment bodies
Similar to what Dr Racket does starting with version 8.4.

The customization variable racket-sexp-comment-fade determines the
amount of fade.

Also: As part of this, I realized racket-font-lock-keywords was broken
in terms of supporting font-lack-maximum-decoration levels. Change it
to a list of variable name symbols, as required. Note that sexp
comment font-lock is a final step for all of the levels; it must be
the final step because the whole idea is that it fades faces already
set by normal font-lock.