Handle truncated error message paths on Racket BC; fixes #604
On Racket BC, error messages can show truncated paths when the paths
are too long. The paths are truncated from the start, and begin with a
"..." prefix.

This doesn't work well with Emacs compilation-mode (to visit error
locations) because the path doesn't exist.

So:

1. Have the Emacs front end's compilation regexp ignore "paths"
starting with "...".

2. Furthermore, display srcloc strings IFF they are not already
present in exn-message. Not only does this also address #604 --
because the full srcloc path is not present in exn-message if it is a
long path that was truncated by Racket BC -- it generalizes our
handling of srclocs. We don't need to test for certain exceptions that
are known to duplicate the first srcloc or all srclocs in exn-message.