Directly break the repl thread
The command to cause a break in the REPL does not need to break-thread
the session/manager thread, for that in turn to break the repl thread.
Instead we can just send the session/manager thread a message telling
it to break-thread the repl thread.

Although this arose from me investigating issue #608, it's orthogonal.
In reviewing the code I just think this new approach is simpler and
clearer. The fix for #608 per se will be its own, next commit.