Add persistent user breakpoints; closes #614
In some sense, especially in the back end, this is a generalization of
racket-debug-run-to-here, which is one, temporary breakpoint.

Each breakpoint has a condition plus a list of actions. The
combination of the condition "#t" and the actions "(break)" is a
simple unconditional breakpoint.

Conditions may use local variable bindings existing around the
breakpoint.

An action list like "(print)" or "(log)" -- not including "break" --
is effectively what some debuggers call a watchpoint instead of a
breakpoint. "print" prints to REPL output, "log" posts to a
racket-mode-debug logger topic. In contrast to the information
supplied to the front end to show "in situ" locals for `break`
actions, we show syntax objects for the watchpoint and locals symbols,
as this might better convey the watchpoint location as well as
disambiguate variable names.