Change compilation-mode regexp; fixes #616
The old regexp exhibited quadratic behavior when encountering certain
text like the JSON example in the bug report.

Simplify and refine the regexp.

Add tests.

Also: A recent commit for #604 had added to the regexp an attempt to
avoid matching filenames starting with "...", so that compilation-mode
would completely ignore them. Although a nice idea, the benefit is
limited to very long filenames only on Racket BC, and it complicates
the regexp. This commit removes that. Instead the "filename adjusting"
function detects this and returns "*unbound*" as the filename (note
that returning nil is allowed in older Emacs, but not newer versions).
Because the filename is not completely ignored, this means the user
could get prompted to enter one, which is not as nice a UX. But again
this is limited to very long filenames on Racket BC, and I think it's
more important to have a simpler, correct regexp for the general case.