Better support for running org-src buffers; fixes #630
Commit 7bcedcc worked OK for an initial run of the buffer, but not
subsequent ones. I failed to fully understand what org-src is doing
with these buffers: It sets a write-contents-functions hook. The hook
copies contents back to the parent org buffer. Furthermore the hook
returns non-nil, meaning no further hooks are run and save-buffer does
not actually write to file.

To accommodate that, as well as our own need to run from an existing
file, we prepend our own write-contents-functions. Our hook really
writes to the temp file. Furthermore our hook returns nil so that the
org-src hook will run, too.

Furthermore, this commit limits this special behavior to when
org-src-mode is enabled in the buffer. Otherwise, we just let
buffer-file-name remain nil, and let racket--save-if-changed do the
usual Emacs thing via save-buffer -- which in the absence of special
write-XXX-functions hooks, prompts the user to save to a real file
using a name of the user's choosing.