Improve support for \\<keymap> in doc strings
When generating reference.org from doc strings, handle in a single new
function (modeled after one in the "helpful" package) all of these
directives:

  \\<keymap>
  \\{keymap}
  \\[command]
  \\=
  conversion of quoted items to refs or code formatting

The main motivation is that \\<keymap> is stateful: it sets the keymap
to use for subsequent uses of \\[command], and may occur multiple
times. Although we don't currently need that (all our doc strings just
use one \\<keymap>) it's good to support for the future.

As a bonus it's more efficient to handle everything in one function
that does one pass through the doc string (as opposed to our old
approach of multiple passes).