Use parse-partial-sexp instead of scan-lists
Commit db7baeb related to #632 added a check for a complete
s-expression using scan-lists before attempting to read. The idea was
this would be less-wasteful than read doing potentially lots of work
and allocations, only to discover end-of-buffer and fail.

However doing scan-lists from the start of the buffer every time is
still more work than necessary. Instead we can use
parse-partial-sexp (asking it to stop when the nesting level reaches
0), which returns state we can use to resume efficiently from where
it last stopped.