Improve configure-runtime and module path handling; fixes #645
Do configure-runtime before namespace-require of user module(s). As a
result, the REPL will use the latter's values for parameters like
print-as-expression.

Furthermore, review handling of configure-runtime and
namespace-require in Racket command-line, and closely follow its
example -- especially with respect to use of module-path-index-join.
This seems to avoid the duplicate Typed Racket error messages (which
was the motivation in commit 3ed2ac6 for moving configure-runtime
later). IIUC type-checking happened twice because the module paths
weren't recognized as equivalent? Although I'm not 100% confident that
is the (entire) explanation, the behavior is correct in testing. In
any case it seems good to follow more closely the example of
command-line Racket itself.