Discard subsequent lines of input after exn:fail:read
This handles cases like #646 where reading "\?"\n errors on the bad \?
escape sequence, leaving the "\n unconsumed from the input port. This
seems like a reasonable mitigation, although I'm not 100% confident
this won't introduce new problems.

[A potentially more reliable (but more drastic) change would be to stop
having the REPL consume interactions from an input port that is
directly connected to the Emacs REPL buffer -- instead, submit to the
back end as a "REPL interaction" command. This would make explicit the
extent of each interaction's input, rather than consuming all pending
lines. There could also be a UI similar to Dr Racket's, for when the
user's program is reading input. But for now, I think all that would
be a solution in search of a problem that few people are
experiencing.]

Also change some `cond` and `match` expressions to sequential uses of
`when` with side effects, to be clearer what's really going on.