racket-{describe company}: Handle non-Scribble docs; fixes #649
Our effort to extract just the maincolumn/main div elements was
failing for non-Scribble-generated docs like those for r5rs. Handle
those for both racket-describe (full doc pages) and
racket-company (for when the user presses C-h and we supply a
"peephole" buffer into only the relevant elements).

Also I realized that an HTML `base` element should really only appear
inside a `head` element. Although it might not matter for the Emacs
shr renderer, clean this up. Rewrite in terms of a function that
returns an entire `body` element. Move the `base` addition up higher,
and put it in the `head` element.