racket-pdb-mode: Don't await pdb-available? response
This can take a few seconds on first load.

Instead, optimistically enable racket-pdb-mode and run pdb-available?
async. When the response arrives, if false, then we can disable
racket-pdb-mode and give show user-error.

This definitely fixes a delay I was able to experience (although only
for the first racket-pdb-mode after restarting Emacs; not thereafter).
Will wait to hear from samth if it fixes the delay they are
experiencing.