racket--modules-at-point: Handle user-error as well as scan-error
This is for when racket-hash-lang-mode is enabled for a non-sexp lang,
and you do C-c C-c for racket-run-module-at-point. It fails to run
with the error message, "cannot move backward". With this commit, it
just always runs the outermost file module with no error message.

A better way to support the idea of "the module at point" would be to
have drracket/check-syntax expose the position and span of module
forms in fully expanded syntax. Which we could gather up in an
interval-map to handle nesting (at any given point, what is the list
of submodules). And then the racket-run-module-at-point command could
use this.

(Such an interval-map of submodule paths is probably related to a
smarter way of supplying completion candidates from imports, such that
only valid ones are supplied for any given point.)