Sync with commit 410deb3 from pdb repo
Use new and renamed get-point-info values.

The new values are an experiment with "semantic" (as opposed to
syntactic) font-lock. This is mainly intended for the day when
racket-pdb-mode might be used alongside a racket-hash-lang-mode, which
does only limited token-based font-lock.

In this experiment, note that the faces disappear briefly during point
motion, as with the other face overlays for things related to point.
Although this is weird, and should eventually change... for now I find
it helpful because it makes obvious which faces are being contributed
by racket-pdb-mode.

Also: Add a check in racket--pdb-on-motion-idle-timer that the window
is still hosting the racket-pdb-mode buffer. It might have switched to
some other buffer (which I experienced when using magit which borrowed
the window for awhile).