racket-show-pseudo-tooltip: Show after end of same line if possible
Show the tooltip after end of the same line if there is room to do so.
Otherwise follow the previous strategies, including trying to show on
the next line (which, although most similar to a GUI tooltip, has a
tendency to obscure something interesting on the next line -- unlike
GUI tooltips, ours don't automatically disappear because that can be
its own annoyance).