Add hook for module language to racket-hash-lang-mode
This is roughly in the same spirit as PR #661, but using the
module-language key supported by new lang's info function.

Also this defines a hook as the means for users to customize. A
default hook function sets comment-start for a few popular
langs (although ultimately comment-start should become a new
"official" lang info key).