Use "ext:" prefix on paredit for declare-function
When packages are not installed in a default location like
~/.emacs.d/elpa then check-declare-{file directory} will fail if it
can't find the file -- unless "ext:" is used.

This is relevant because on my personal setup I recently moved my
packages dir. But Makefile runs emacs -Q and doesn't get my
package-user-dir value. For now, just ignore this when running
locally; the check still runs on CI.