Add racket-hash-lang-token-face-alist
Instead of hard-coding the mapping of tokens to faces, allow the user
to customize most of them.

(If the set of tokens were closed, it would suffice to define faces
for the user to customize. Which was my initial direction. But since
the set of tokens is open, and we're letting the user map any token
symbol to any face, we don't really need those faces anymore, so
delete them.)

While adding racket-hash-lang-token-face-alist to the reference docs,
also add some of the racket-hash-lang-mode commands and keybinds.