Fix cl-loop byte-compiler warnings
Frustratingly, the work-around of introducing a dummy `let` binding
for the cl-loop index variable works on older versions of Emacs... but
on bleeding edge Emacs starting c. a week ago starts to cause a
different warning unless the variable is used.

So to make this work across all versions (so far) of Emacs, we need a
work-around for the work-around.