racket-xp-mode: Do "semantic" highlighting of binding sites
This is mainly intended for, and by default only enabled for, the
major mode racket-hash-lang-mode.

This enables a "more colorful" approach, which is more similar to
"classic" racket-mode -- although it is not, and not intended to be,
exactly the same.

So the basic story here is: racket-hash-lang-mode does a basic level
of fontification applying 'face properties as usual, based on lexer
tokens. Then when racket-xp-mode is enabled, and this extra
highlighting is enabled, it will add `font-lock-face properties. Note
that 'face always wins over 'font-lock-face, so the minor mode can
never override the basic token coloring decisions, but it can enhance
text that lacks any 'face property at all. By configuring
hash-lang-mode not to apply any 'face for "symbol" tokens -- typically
the vast majority of a source -- this leaves the field free for a
minor mode like racket-xp-mode to augment, if a user prefers.