Make racket-repl-mode even plainer
A previous commit set font-lock-defaults to nil, but neglected to do
so for syntax-propertize-function. Moreover the default syntax-table
really assumes too much about the lang being used, as well as causing
issues like #633.

This commit leaves racket-repl-mode looking _very_ plain and with no
expression navigability. Which is actually desirable for e.g.
arbitrary user program output. The right way to restore some of this
is via some separation of output elements as discussed in #633 as well
as in #667. But meanwhile this commit is definitely in the spirit of
making things simpler, to be correct, before correctly remaking them
to be fancier again. :(

We need to disable some racket-repl-mode tests related to multi-line
expression input and paredit, which can no longer work. I'm commenting
them out with the idea they could be revived when we someday revive
fancier input,