Revive hash-lang in racket-repl-mode
Because a run command may be issued from any edit buffer -- using
either racket-mode or racket-hash-lang-mode -- it should configure the
REPL appropriately.

A previous commit had:

  - changed racket-hash-lang-mode from a minor mode to a major mode,
  making it unable to enhance racket-repl-mode anymore

  - changed the REPL to font-lock nothing and use very plain
  char-syntax for everything, including user input and program output

Although that last change is still desirable for _output_ (see e.g.
issue #633), this commit revives using hash-lang stuff for _input_.

I'm not 100% confident in this commit, but want to at least preserve a
check-point before testing more and potentially exploring other
approaches. One tricky point was precisely _when_ to config the REPL
buffer. For example the user may issue a run command when the REPL
buffer doesn't yet exist; it gets created and displayed as an effect
of the run command. So a "before-run" hook is too early. OTOH an
"after-run" hook might be too late, because someday we might want to
use a hash-lang to format `current-print`ed values (if not other
current-output-port junk). For now I'm defining a distinct new hook
for this; although it's just run with the other "after" hooks, today,
in theory its timing is a warm bowl of porridge that could be adjusted
later.