Better manage back end hash-lang object state/lifetimes
Create the back end hash-lang object using an async command, so as not
to block Emacs. (If the back end isn't already running, and the
command causes it to start, this can take awhile.) Limit the buffer --
read-only, no font-lock -- until the command response says the object
is ready for use.

Handle the back end server being stopped -- which yanks the rug out
from under racket-hash-lang-mode -- as well as being later restarted.
When the back end stops, "downgrade" buffers to plain racket-mode, and
add them to a list for restoration to racket-hash-lang-mode when the
back end is later restarted. Similarly put any associated
racket-repl-mode buffers into a state as if the most recent run
command had been from a racket-mode buffer.

Although I can't so far reproduce #669, this commit _might_ fix that
because it avoids some conditions under which that could occur.