racket-hash-lang-mode: Use set-font-lock-defaults; fixes #642
Credit and big thanks to @dpk for identifying the root cause!

I had been unable to reproduce the problem because of my use of
paren-face-mode; it was doing a font-lock-ensure, which in turn calls
font-lock-set-defaults. It suffices for us to call the latter
directly.

Also:

- Group the font-lock-xxx items at the start of the major mode
function. Move some "miscellaneous" settings to the end.

- Remove redundant init of racket--hash-lang-generation; this is
leftover from it being a minor mode, and furthermore it's now the
responsibility of racket--hash-lang-{create delete}.

- Update doc string about how much coloring to expect, and a hint that
racket-xp-mode can contribute more colors.