Rely on wrapper of edit mode's font-lock-fontify-region
Instead of worrying so much about the buffer's syntax-table, and text
properties setting such, focus on the fontification: Use the edit
mode's font-lock-fontify-region-function value, just in a wrapper that
limits it to the input tail of the buffer. Similarly have the
racket-hash-lang-repl-buffer-string function work in those terms, as
opposed to looking for 'field properties.