Handle input for user programs
Allow buffer tail >= pmark to be read/write even when not immediately
following a REPL prompt.

When at a prompt, handle input as REPL expression submission as
before (including adding to REPL expression input history).

Otherwise treat it as input for user program.

Similarly, our fontify-region wrapper continues to fontify richly when
following a prompt, for REPL expression input -- but not for the
arbitrary user program input that we now allow.