Refine output and prompt handling
Use two markers: One for the single live REPL interaction prompt, if
any, always toward the end of the buffer, and another for user program
I/O and back end output messages. The latter is roughly like
process-mark, and previous commits' racket--repl-pmark variable, with
the clarification that its role is more limited.

When the REPL has a live prompt, always insert program output /before/
the prompt, pushing it down. In other words, the output mark is always
less than the REPL mark if any. This handles scenarios like when a
user program creates threads that continue to output even when we're
back at a REPL prompt-read.

Also: Carefully review the read-only and rear-nonsticky property
handling so that there are no read/write "gaps" left in output or in old
input that we intentionally want to make all read-only.